copy updated files from gitee to github

+++ built-in/PyTorch/Official/cv/image_object_detection/deeplabv3+_ID0326_for_PyTorch/utils/loss.py
@@ -58,7 +58,8 @@ class SegmentationLosses(object):
         if self.npu:
             criterion = criterion.npu()
 
-        loss = criterion(logit, target.long())
+        #loss = criterion(logit, target.long())
+        loss = criterion(logit, target.to(torch.int32))
 
         if self.batch_average:
             loss /= n
